CI: move components to _components folder
