+++ ci/src/app/_components/home/home.component.ts
@@ -1,6 +1,6 @@
 ﻿import { Component, OnInit } from '@angular/core';
 
-import { TopicService } from '../_services/index';
+import { TopicService } from '../../_services/index';
 
 @Component({
     moduleId: module.id,
