+++ ci/src/app/_components/topic/topic.component.ts
@@ -1,6 +1,6 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { TopicService } from '../_services/index';
+import { TopicService } from '../../_services/index';
 
 @Component({
     moduleId: module.id,
