+++ ci/src/app/_components/welcome/register/index.ts
