+++ ci/src/app/_components/welcome/register/register.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthenticationService } from '../_services/index';
+import { AuthenticationService } from '../../../_services/index';
 
 @Component({
     moduleId: module.id,
