+++ ci/src/app/_components/welcome/welcome.component.ts
@@ -1,14 +1,14 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthenticationService } from '../_services/index';
+import { AuthenticationService } from '../../_services/index';
 
 @Component({
     moduleId: module.id,
-    templateUrl: 'login.component.html'
+    templateUrl: 'welcome.component.html'
 })
 
-export class LoginComponent implements OnInit {
+export class WelcomeComponent implements OnInit {
     model: any = {};
     loading = false;
     error = '';
